Replacing old gif icons with new high quality png icons
Since the current icons are still in gif format,
they look pretty bad on a dark theme in Eclipse.

So adding new replacement icons, which are exported as png,
which makes a much better quality and looks good on a dark theme.
Icon size is 16x16, as toolbar icons have to be that size format.

Also adding the svg (source) icons, which was the source for the
png variant. This makes it possible to create a different icon size,
if Eclipse would change that.

Closes: https://github.com/tomcatplugin/tomcatplugin/issues/10
Signed-off-by: Conrad Kostecki <conrad@kostecki.com>