Replacing more old gif icons with new high quality png icons
As there are more old icons left in gif format,
it's time to update them too :-)

So adding new replacement icons, which are exported as png,
which makes a much better quality and looks good on a dark theme.

Also adding the svg (source) icons, which was the source for the
png variant. This makes it possible to create a different icon size,
if Eclipse would change that.

Signed-off-by: Conrad Kostecki <conrad@kostecki.com>